Many bugfixes
Implemented resetStatus and resetPieceToMove functions to be called when
undo is clicked so that the gameStatus returns to "" and the pieceToMove
is back to null (and hence no piece can be moved); implemented setStatus
method that is called when the game is over (goes through each winner
and adds to the gameStatus string). Implemented resetGame to be called
when a new game is loaded from the disk; reinstantiates the boardHistory
and sets the board's colors to those passed in from the file. Edited
nextTurn() so that if the gameStatus indicates a winner then nextTurn
will not do anything. Created a helper method to find the winners
(findWinners): goes through each player and finds the lowest scores,
adding the numbers of the players to whom those scores belong to into a
LinkedList and passing it to the setStatus method.